+++ csfieldguide/chapters/models.py
@@ -38,10 +38,3 @@ class Chapter(models.Model):
         """
         return self.name
 
-
-class Interactive(models.Model):
-    """Model for interactive in database."""
-
-    #  Auto-incrementing 'id' field is automatically set by Django
-    slug = models.SlugField(unique=True)
-    template = models.CharField(max_length=150)
