+++ csfieldguide/chapters/urls.py
@@ -12,12 +12,6 @@ urlpatterns = [
         views.IndexView.as_view(),
         name="home"
     ),
-    # eg: /chapters/interactives/sorting-algorithms/
-    url(
-        r"^interactives/(?P<interactive_slug>[-\w]+)/$",
-        views.InteractiveView.as_view(),
-        name="interactive"
-    ),
     # eg: /chapters/glossary/
     url(
         r"^glossary/$",
