+++ csfieldguide/chapters/views.py
@@ -6,7 +6,7 @@ from django.http import JsonResponse, Http404
 
 from general.templatetags.render_html_field import render_html_with_static
 
-from .models import Chapter, GlossaryTerm, Interactive
+from .models import Chapter, GlossaryTerm
 
 
 class IndexView(generic.ListView):
@@ -45,20 +45,6 @@ class ChapterView(generic.DetailView):
     model = Chapter
     template_name = "chapters/chapter.html"
     slug_url_kwarg = "chapter_slug"
-
-
-class InteractiveView(generic.DetailView):
-    """View for a specific topic."""
-
-    model = Interactive
-    slug_url_kwarg = "interactive_slug"
-    template_name = "chapters/interactives/whole-page-interactive-base.html"
-    # context_object_name = "interactive"
-
-    # def get_context_data(self, **kwargs):
-    #     context["interactive_template"] = "chapters/interactives/{}.html".format(
-    #         self.object.slug
-    #     )
 
 
 def glossary_json(request, **kwargs):
