+++ csfieldguide/config/settings/base.py
@@ -38,6 +38,7 @@ DJANGO_APPS = [
 # Apps specific for this project go here.
 LOCAL_APPS = [
     "general.apps.GeneralConfig",
+    "interactives.apps.InteractivesConfig",
     "chapters.apps.ChaptersConfig",
 ]
