+++ csfieldguide/general/management/commands/updatedata.py
@@ -11,4 +11,5 @@ class Command(management.base.BaseCommand):
     def handle(self, *args, **options):
         """Automatically called when the updatedata command is given."""
         management.call_command("flush", interactive=False)
+        management.call_command("loadinteractives")
         management.call_command("loadchapters")
