+++ csfieldguide/interactives/apps.py
@@ -1,0 +1,9 @@
+"""Application configuration for the interactives application."""
+
+from django.apps import AppConfig
+
+
+class InteractivesConfig(AppConfig):
+    """Configuration object for the interactives application."""
+    
+    name = "interactives"
